requirements.txt
pdfplumber
python-docx
google-genai
openai
